@@ -11,7 +11,7 @@
   ],
   "scripts": {
     "build": "fedx-scripts webpack",
-    "i18n_extract": "BABEL_ENV=i18n fedx-scripts babel src --quiet > /dev/null",
+    "i18n_extract": "fedx-scripts formatjs extract",
     "lint": "fedx-scripts eslint --ext .js --ext .jsx .",
     "snapshot": "fedx-scripts jest --updateSnapshot",
     "start": "fedx-scripts webpack-dev-server --progress",
@@ -32,57 +32,54 @@
     "url": "https://github.com/openedx/frontend-app-authn/issues"
   },
   "dependencies": {
-    "@edx/brand": "npm:@edx/brand-openedx@1.1.0",
-    "@edx/frontend-component-cookie-policy-banner": "2.2.0",
-    "@edx/frontend-platform": "4.2.0",
-    "@edx/paragon": "20.20.0",
-    "@fortawesome/fontawesome-svg-core": "6.2.1",
-    "@fortawesome/free-brands-svg-icons": "6.2.1",
-    "@fortawesome/free-regular-svg-icons": "6.2.1",
-    "@fortawesome/free-solid-svg-icons": "6.2.1",
-    "@fortawesome/react-fontawesome": "0.2.0",
-    "@redux-devtools/extension": "3.2.3",
-    "classnames": "2.3.2",
-    "core-js": "3.26.1",
-    "extract-react-intl-messages": "4.1.1",
+    "@edx/brand": "npm:@openedx/brand-openedx@^1.2.2",
+    "@edx/frontend-platform": "^8.0.0",
+    "@edx/openedx-atlas": "^0.6.0",
+    "@fortawesome/fontawesome-svg-core": "6.6.0",
+    "@fortawesome/free-brands-svg-icons": "6.6.0",
+    "@fortawesome/free-solid-svg-icons": "6.6.0",
+    "@fortawesome/react-fontawesome": "0.2.2",
+    "@openedx/paragon": "^22.1.1",
+    "@optimizely/react-sdk": "^2.9.1",
+    "@redux-devtools/extension": "3.3.0",
+    "@testing-library/react": "^12.1.5",
+    "@testing-library/react-hooks": "^8.0.1",
+    "algoliasearch": "^4.14.3",
+    "algoliasearch-helper": "^3.14.0",
+    "classnames": "2.5.1",
+    "core-js": "3.38.1",
     "fastest-levenshtein": "1.0.16",
-    "form-urlencoded": "4.2.1",
-    "lodash.camelcase": "4.3.0",
-    "lodash.snakecase": "4.1.1",
+    "form-urlencoded": "6.1.5",
     "prop-types": "15.8.1",
-    "query-string": "5.1.1",
-    "react": "16.14.0",
-    "react-dom": "16.14.0",
+    "query-string": "7.1.3",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
     "react-helmet": "6.1.0",
-    "react-loading-skeleton": "2.2.0",
-    "react-onclickoutside": "6.12.2",
+    "react-loading-skeleton": "3.5.0",
     "react-redux": "7.2.9",
     "react-responsive": "8.2.0",
-    "react-router": "5.3.4",
-    "react-router-dom": "5.3.4",
-    "redux": "4.2.0",
+    "react-router": "6.26.2",
+    "react-router-dom": "6.26.2",
+    "react-zendesk": "^0.1.13",
+    "redux": "4.2.1",
     "redux-logger": "3.0.6",
     "redux-mock-store": "1.5.4",
-    "redux-saga": "1.2.1",
+    "redux-saga": "1.3.0",
     "redux-thunk": "2.4.2",
-    "regenerator-runtime": "0.13.11",
-    "reselect": "4.1.7",
-    "sanitize-html": "2.7.3",
-    "semver-regex": "3.1.4",
+    "regenerator-runtime": "0.14.1",
+    "reselect": "4.1.8",
     "universal-cookie": "4.0.4"
   },
   "devDependencies": {
     "@edx/browserslist-config": "^1.1.1",
-    "@edx/frontend-build": "11.0.2",
     "@edx/reactifex": "1.1.0",
-    "babel-plugin-formatjs": "10.3.31",
-    "enzyme": "3.11.0",
-    "enzyme-adapter-react-16": "1.15.7",
-    "eslint-plugin-import": "2.26.0",
+    "@openedx/frontend-build": "^14.0.3",
+    "babel-plugin-formatjs": "10.5.16",
+    "eslint-plugin-import": "2.30.0",
     "glob": "7.2.3",
     "history": "5.3.0",
     "husky": "7.0.4",
-    "jest": "27.5.1",
-    "react-test-renderer": "16.14.0"
+    "jest": "29.7.0",
+    "react-test-renderer": "^17.0.2"
   }
 }
