@@ -1,29 +1,13 @@
 import { defineMessages } from '@edx/frontend-platform/i18n';
 
 const messages = defineMessages({
+  // institution login strings
   'institution.login.page.sub.heading': {
     id: 'institution.login.page.sub.heading',
     defaultMessage: 'Choose your institution from the list below',
     description: 'Heading of the institutions list',
   },
-  // Confirmation Alert Message
-  'forgot.password.confirmation.title': {
-    id: 'forgot.password.confirmation.title',
-    defaultMessage: 'Check your email',
-    description: 'Forgot password confirmation message title',
-  },
-  'forgot.password.confirmation.support.link': {
-    id: 'forgot.password.confirmation.support.link',
-    defaultMessage: 'contact technical support',
-    description: 'Technical support link text',
-  },
-  'forgot.password.confirmation.info': {
-    id: 'forgot.password.confirmation.info',
-    defaultMessage: 'If you do not receive a password reset message after 1 minute, verify that you entered the correct '
-                    + 'email address, or check your spam folder.',
-    description: 'Part of message that appears after user requests password change',
-  },
-  // Logistration strinsg
+  // logistration strings
   'logistration.sign.in': {
     id: 'logistration.sign.in',
     defaultMessage: 'Sign in',
@@ -34,26 +18,11 @@ const messages = defineMessages({
     defaultMessage: 'Register',
     description: 'Text that appears on the tab to switch between login and register',
   },
-  'internal.server.error.message': {
-    id: 'internal.server.error.message',
-    defaultMessage: 'An error has occurred. Try refreshing the page, or check your internet connection.',
-    description: 'Error message that appears when server responds with 500 error code',
-  },
-  'server.ratelimit.error.message': {
-    id: 'server.ratelimit.error.message',
-    defaultMessage: 'An error has occurred because of too many requests. Please try again after some time.',
-    description: 'Error message that appears when server responds with 429 error code',
-  },
   // enterprise sso strings
   'enterprisetpa.title.heading': {
     id: 'enterprisetpa.title.heading',
     defaultMessage: 'Would you like to sign in using your {providerName} credentials?',
     description: 'Header text used in enterprise third party authentication',
-  },
-  'enterprisetpa.sso.button.title': {
-    id: 'enterprisetpa.sso.button.title',
-    defaultMessage: 'Sign in using {providerName}',
-    description: 'Text for third party auth provider buttons',
   },
   'enterprisetpa.login.button.text': {
     id: 'enterprisetpa.login.button.text',
@@ -127,6 +96,41 @@ const messages = defineMessages({
                     + 'before you start learning with {platformName}.',
     description: 'Message that appears on register page if user has successfully authenticated with TPA '
                   + 'but no associated platform account exists',
+  },
+  'registration.using.tpa.form.heading': {
+    id: 'registration.using.tpa.form.heading',
+    defaultMessage: 'Finish creating your account',
+    description: 'Heading that appears above form when user is trying to create account using social auth',
+  },
+  supportTitle: {
+    id: 'zendesk.supportTitle',
+    description: 'Title for the support button',
+    defaultMessage: 'edX Support',
+  },
+  selectTicketForm: {
+    id: 'zendesk.selectTicketForm',
+    description: 'Select ticket form',
+    defaultMessage: 'Please choose your request type:',
+  },
+  'registration.other.options.heading': {
+    id: 'registration.other.options.heading',
+    defaultMessage: 'Or register with:',
+    description: 'A message that appears above third party auth providers i.e saml, google, facebook etc',
+  },
+  'institution.login.button': {
+    id: 'institution.login.button',
+    defaultMessage: 'Institution/campus credentials',
+    description: 'shows institutions list',
+  },
+  'login.other.options.heading': {
+    id: 'login.other.options.heading',
+    defaultMessage: 'Or sign in with:',
+    description: 'Text that appears above other sign in options like social auth buttons',
+  },
+  'enterprise.login.btn.text': {
+    id: 'enterprise.login.btn.text',
+    defaultMessage: 'Company or school credentials',
+    description: 'Company or school login link text.',
   },
 });
 
