@@ -3,9 +3,9 @@ import { all } from 'redux-saga/effects';
 import { saga as commonComponentsSaga } from '../common-components';
 import { saga as forgotPasswordSaga } from '../forgot-password';
 import { saga as loginSaga } from '../login';
+import { saga as authnProgressiveProfilingSaga } from '../progressive-profiling';
 import { saga as registrationSaga } from '../register';
 import { saga as resetPasswordSaga } from '../reset-password';
-import { saga as welcomePageSaga } from '../welcome';
 
 export default function* rootSaga() {
   yield all([
@@ -14,6 +14,6 @@ export default function* rootSaga() {
     commonComponentsSaga(),
     forgotPasswordSaga(),
     resetPasswordSaga(),
-    welcomePageSaga(),
+    authnProgressiveProfilingSaga(),
   ]);
 }
