@@ -1,10 +1,11 @@
 export {
   getTpaProvider,
   getTpaHint,
-  updatePathWithQueryParams,
-  getAllPossibleQueryParam,
+  getAllPossibleQueryParams,
   getActivationStatus,
+  isHostAvailableInQueryParams,
+  updatePathWithQueryParams,
   windowScrollTo,
 } from './dataUtils';
 export { default as AsyncActionType } from './reduxUtils';
-export { default as setSurveyCookie, setCookie } from './cookies';
+export { default as setCookie } from './cookies';
