@@ -1,13 +1,13 @@
 import React from 'react';
 
-import { Form, Icon } from '@edx/paragon';
-import { ExpandMore } from '@edx/paragon/icons';
+import { Form, Icon } from '@openedx/paragon';
+import { ExpandMore } from '@openedx/paragon/icons';
 import PropTypes from 'prop-types';
 
 const FormFieldRenderer = (props) => {
   let formField = null;
   const {
-    errorMessage, fieldData, onChangeHandler, isRequired, value,
+    className, errorMessage, fieldData, onChangeHandler, isRequired, value,
   } = props;
 
   const handleFocus = (e) => {
@@ -26,6 +26,7 @@ const FormFieldRenderer = (props) => {
       formField = (
         <Form.Group controlId={fieldData.name} isInvalid={!!(isRequired && errorMessage)}>
           <Form.Control
+            className={className}
             as="select"
             name={fieldData.name}
             value={value}
@@ -54,6 +55,7 @@ const FormFieldRenderer = (props) => {
       formField = (
         <Form.Group controlId={fieldData.name} isInvalid={!!(isRequired && errorMessage)}>
           <Form.Control
+            className={className}
             as="textarea"
             name={fieldData.name}
             value={value}
@@ -76,6 +78,7 @@ const FormFieldRenderer = (props) => {
       formField = (
         <Form.Group controlId={fieldData.name} isInvalid={!!(isRequired && errorMessage)}>
           <Form.Control
+            className={className}
             name={fieldData.name}
             value={value}
             aria-invalid={isRequired && Boolean(errorMessage)}
@@ -97,6 +100,7 @@ const FormFieldRenderer = (props) => {
       formField = (
         <Form.Group isInvalid={!!(isRequired && errorMessage)}>
           <Form.Checkbox
+            className={className}
             id={fieldData.name}
             checked={!!value}
             name={fieldData.name}
@@ -124,6 +128,7 @@ const FormFieldRenderer = (props) => {
   return formField;
 };
 FormFieldRenderer.defaultProps = {
+  className: '',
   value: '',
   handleBlur: null,
   handleFocus: null,
@@ -132,17 +137,22 @@ FormFieldRenderer.defaultProps = {
 };
 
 FormFieldRenderer.propTypes = {
+  className: PropTypes.string,
   fieldData: PropTypes.shape({
     type: PropTypes.string,
     label: PropTypes.string,
     name: PropTypes.string,
+    options: PropTypes.arrayOf(PropTypes.arrayOf(PropTypes.string)),
   }).isRequired,
   onChangeHandler: PropTypes.func.isRequired,
   handleBlur: PropTypes.func,
   handleFocus: PropTypes.func,
   errorMessage: PropTypes.string,
   isRequired: PropTypes.bool,
-  value: PropTypes.string,
+  value: PropTypes.oneOfType([
+    PropTypes.string,
+    PropTypes.bool,
+  ]),
 };
 
 export default FormFieldRenderer;
