@@ -1,1 +1,2 @@
-export { default } from './FieldRenderer';
+/* eslint-disable import/prefer-default-export */
+export { default as FormFieldRenderer } from './FieldRenderer';
