@@ -3,7 +3,7 @@ import { connect } from 'react-redux';
 
 import { getConfig } from '@edx/frontend-platform';
 import { sendPageEvent, sendTrackEvent } from '@edx/frontend-platform/analytics';
-import { injectIntl, intlShape } from '@edx/frontend-platform/i18n';
+import { useIntl } from '@edx/frontend-platform/i18n';
 import {
   Form,
   Hyperlink,
@@ -11,33 +11,34 @@ import {
   StatefulButton,
   Tab,
   Tabs,
-} from '@edx/paragon';
-import { ChevronLeft } from '@edx/paragon/icons';
+} from '@openedx/paragon';
+import { ChevronLeft } from '@openedx/paragon/icons';
 import PropTypes from 'prop-types';
 import { Helmet } from 'react-helmet';
-import { Redirect } from 'react-router-dom';
+import { useNavigate } from 'react-router-dom';
 
-import BaseComponent from '../base-component';
-import { FormGroup } from '../common-components';
-import { DEFAULT_STATE, LOGIN_PAGE, VALID_EMAIL_REGEX } from '../data/constants';
-import { updatePathWithQueryParams, windowScrollTo } from '../data/utils';
 import { forgotPassword, setForgotPasswordFormData } from './data/actions';
 import { forgotPasswordResultSelector } from './data/selectors';
 import ForgotPasswordAlert from './ForgotPasswordAlert';
 import messages from './messages';
+import BaseContainer from '../base-container';
+import { FormGroup } from '../common-components';
+import { DEFAULT_STATE, LOGIN_PAGE, VALID_EMAIL_REGEX } from '../data/constants';
+import { updatePathWithQueryParams, windowScrollTo } from '../data/utils';
 
 const ForgotPasswordPage = (props) => {
   const platformName = getConfig().SITE_NAME;
   const emailRegex = new RegExp(VALID_EMAIL_REGEX, 'i');
   const {
-    intl, status, submitState, emailValidationError,
+    status, submitState, emailValidationError,
   } = props;
 
+  const { formatMessage } = useIntl();
   const [email, setEmail] = useState(props.email);
   const [bannerEmail, setBannerEmail] = useState('');
   const [formErrors, setFormErrors] = useState('');
   const [validationError, setValidationError] = useState(emailValidationError);
-  const [key, setKey] = useState('');
+  const navigate = useNavigate();
 
   useEffect(() => {
     sendPageEvent('login_and_registration', 'reset');
@@ -58,9 +59,9 @@ const ForgotPasswordPage = (props) => {
     let error = '';
 
     if (value === '') {
-      error = intl.formatMessage(messages['forgot.password.empty.email.field.error']);
+      error = formatMessage(messages['forgot.password.empty.email.field.error']);
     } else if (!emailRegex.test(value)) {
-      error = intl.formatMessage(messages['forgot.password.page.invalid.email.message']);
+      error = formatMessage(messages['forgot.password.page.invalid.email.message']);
     }
 
     return error;
@@ -89,35 +90,32 @@ const ForgotPasswordPage = (props) => {
   const tabTitle = (
     <div className="d-inline-flex flex-wrap align-items-center">
       <Icon src={ChevronLeft} />
-      <span className="ml-2">{intl.formatMessage(messages['sign.in.text'])}</span>
+      <span className="ml-2">{formatMessage(messages['sign.in.text'])}</span>
     </div>
   );
 
   return (
-    <BaseComponent>
+    <BaseContainer>
       <Helmet>
-        <title>{intl.formatMessage(messages['forgot.password.page.title'],
+        <title>{formatMessage(messages['forgot.password.page.title'],
           { siteName: getConfig().SITE_NAME })}
         </title>
       </Helmet>
       <div>
-        <Tabs activeKey="" id="controlled-tab" onSelect={(k) => setKey(k)}>
+        <Tabs activeKey="" id="controlled-tab" onSelect={(key) => navigate(updatePathWithQueryParams(key))}>
           <Tab title={tabTitle} eventKey={LOGIN_PAGE} />
         </Tabs>
-        { key && (
-          <Redirect to={updatePathWithQueryParams(key)} />
-        )}
         <div id="main-content" className="main-content">
           <Form id="forget-password-form" name="forget-password-form" className="mw-xs">
             <ForgotPasswordAlert email={bannerEmail} emailError={formErrors} status={status} />
             <h2 className="h4">
-              {intl.formatMessage(messages['forgot.password.page.heading'])}
+              {formatMessage(messages['forgot.password.page.heading'])}
             </h2>
             <p className="mb-4">
-              {intl.formatMessage(messages['forgot.password.page.instructions'])}
+              {formatMessage(messages['forgot.password.page.instructions'])}
             </p>
             <FormGroup
-              floatingLabel={intl.formatMessage(messages['forgot.password.page.email.field.label'])}
+              floatingLabel={formatMessage(messages['forgot.password.page.email.field.label'])}
               name="email"
               value={email}
               autoComplete="on"
@@ -125,17 +123,17 @@ const ForgotPasswordPage = (props) => {
               handleChange={(e) => setEmail(e.target.value)}
               handleBlur={handleBlur}
               handleFocus={handleFocus}
-              helpText={[intl.formatMessage(messages['forgot.password.email.help.text'], { platformName })]}
+              helpText={[formatMessage(messages['forgot.password.email.help.text'], { platformName })]}
             />
             <StatefulButton
               id="submit-forget-password"
               name="submit-forget-password"
               type="submit"
               variant="brand"
-              className="forgot-password-button-width"
+              className="forgot-password--button"
               state={submitState}
               labels={{
-                default: intl.formatMessage(messages['forgot.password.page.submit.button']),
+                default: formatMessage(messages['forgot.password.page.submit.button']),
                 pending: '',
               }}
               onClick={handleSubmit}
@@ -150,11 +148,11 @@ const ForgotPasswordPage = (props) => {
                 target="_blank"
                 showLaunchIcon={false}
               >
-                {intl.formatMessage(messages['need.help.sign.in.text'])}
+                {formatMessage(messages['need.help.sign.in.text'])}
               </Hyperlink>
             )}
             <p className="mt-5.5 small text-gray-700">
-              {intl.formatMessage(messages['additional.help.text'], { platformName })}
+              {formatMessage(messages['additional.help.text'], { platformName })}
               <span>
                 <Hyperlink isInline destination={`mailto:${getConfig().INFO_EMAIL}`}>{getConfig().INFO_EMAIL}</Hyperlink>
               </span>
@@ -162,7 +160,7 @@ const ForgotPasswordPage = (props) => {
           </Form>
         </div>
       </div>
-    </BaseComponent>
+    </BaseContainer>
   );
 };
 
@@ -170,7 +168,6 @@ ForgotPasswordPage.propTypes = {
   email: PropTypes.string,
   emailValidationError: PropTypes.string,
   forgotPassword: PropTypes.func.isRequired,
-  intl: intlShape.isRequired,
   setForgotPasswordFormData: PropTypes.func.isRequired,
   status: PropTypes.string,
   submitState: PropTypes.string,
@@ -189,4 +186,4 @@ export default connect(
     forgotPassword,
     setForgotPasswordFormData,
   },
-)(injectIntl(ForgotPasswordPage));
+)(ForgotPasswordPage);
