@@ -1,15 +1,12 @@
 import React from 'react';
 import { Provider } from 'react-redux';
 
-import CookiePolicyBanner from '@edx/frontend-component-cookie-policy-banner';
 import { mergeConfig } from '@edx/frontend-platform';
-import * as analytics from '@edx/frontend-platform/analytics';
-import * as auth from '@edx/frontend-platform/auth';
 import { configure, injectIntl, IntlProvider } from '@edx/frontend-platform/i18n';
-import { mount } from 'enzyme';
-import { createMemoryHistory } from 'history';
-import { act } from 'react-dom/test-utils';
-import { MemoryRouter, Router } from 'react-router-dom';
+import {
+  fireEvent, render, screen,
+} from '@testing-library/react';
+import { MemoryRouter } from 'react-router-dom';
 import configureStore from 'redux-mock-store';
 
 import { INTERNAL_SERVER_ERROR, LOGIN_PAGE } from '../../data/constants';
@@ -17,14 +14,20 @@ import { PASSWORD_RESET } from '../../reset-password/data/constants';
 import { setForgotPasswordFormData } from '../data/actions';
 import ForgotPasswordPage from '../ForgotPasswordPage';
 
-jest.mock('@edx/frontend-platform/analytics');
-jest.mock('@edx/frontend-platform/auth');
+const mockedNavigator = jest.fn();
 
-analytics.sendPageEvent = jest.fn();
+jest.mock('@edx/frontend-platform/analytics', () => ({
+  sendPageEvent: jest.fn(),
+  sendTrackEvent: jest.fn(),
+}));
+jest.mock('@edx/frontend-platform/auth');
+jest.mock('react-router-dom', () => ({
+  ...(jest.requireActual('react-router-dom')),
+  useNavigate: () => mockedNavigator,
+}));
 
 const IntlForgotPasswordPage = injectIntl(ForgotPasswordPage);
 const mockStore = configureStore();
-const history = createMemoryHistory();
 
 const initialState = {
   forgotPassword: {
@@ -51,7 +54,12 @@ describe('ForgotPasswordPage', () => {
 
   beforeEach(() => {
     store = mockStore(initialState);
-    auth.getAuthenticatedUser = jest.fn(() => ({ userId: 3, username: 'edX' }));
+    jest.mock('@edx/frontend-platform/auth', () => ({
+      getAuthenticatedUser: jest.fn(() => ({
+        userId: 3,
+        username: 'test-user',
+      })),
+    }));
     configure({
       loggingService: { logError: jest.fn() },
       config: {
@@ -65,6 +73,15 @@ describe('ForgotPasswordPage', () => {
       status: null,
     };
   });
+  const findByTextContent = (container, text) => Array.from(container.querySelectorAll('*')).find(
+    element => element.textContent === text,
+  );
+
+  it('not should display need other help signing in button', () => {
+    const { queryByTestId } = render(reduxWrapper(<IntlForgotPasswordPage {...props} />));
+    const forgotPasswordButton = queryByTestId('forgot-password');
+    expect(forgotPasswordButton).toBeNull();
+  });
 
   it('not should display need other help signing in button', () => {
     const wrapper = mount(reduxWrapper(<IntlForgotPasswordPage {...props} />));
@@ -75,21 +92,25 @@ describe('ForgotPasswordPage', () => {
     mergeConfig({
       LOGIN_ISSUE_SUPPORT_LINK: '/support',
     });
-    const wrapper = mount(reduxWrapper(<IntlForgotPasswordPage {...props} />));
-    expect(wrapper.find('#forgot-password').first().text()).toEqual('Need help signing in?');
+    render(reduxWrapper(<IntlForgotPasswordPage {...props} />));
+    const forgotPasswordButton = screen.findByText('Need help signing in?');
+    expect(forgotPasswordButton).toBeDefined();
   });
 
   it('should display email validation error message', async () => {
     const validationMessage = 'We were unable to contact you.Enter a valid email address below.';
-    const wrapper = mount(reduxWrapper(<IntlForgotPasswordPage {...props} />));
+    const { container } = render(reduxWrapper(<IntlForgotPasswordPage {...props} />));
 
-    wrapper.find('input#email').simulate(
-      'change', { target: { value: 'invalid-email', name: 'email' } },
-    );
-    await act(async () => { await wrapper.find('button.btn-brand').simulate('click'); });
-    wrapper.update();
+    const emailInput = screen.getByLabelText('Email');
 
-    expect(wrapper.find('.alert-danger').text()).toEqual(validationMessage);
+    fireEvent.change(emailInput, { target: { value: 'invalid-email' } });
+
+    const submitButton = screen.getByText('Submit');
+    fireEvent.click(submitButton);
+
+    const alertElements = container.querySelectorAll('.alert-danger');
+    const validationErrors = alertElements[0].textContent;
+    expect(validationErrors).toBe(validationMessage);
   });
 
   it('should show alert on server error', () => {
@@ -98,19 +119,25 @@ describe('ForgotPasswordPage', () => {
     });
     const expectedMessage = 'We were unable to contact you.'
                             + 'An error has occurred. Try refreshing the page, or check your internet connection.';
-    const wrapper = mount(reduxWrapper(<IntlForgotPasswordPage {...props} />));
 
-    expect(wrapper.find('#validation-errors').first().text()).toEqual(expectedMessage);
+    const { container } = render(reduxWrapper(<IntlForgotPasswordPage {...props} />));
+
+    const alertElements = container.querySelectorAll('.alert-danger');
+    const validationErrors = alertElements[0].textContent;
+    expect(validationErrors).toBe(expectedMessage);
   });
 
   it('should display empty email validation message', async () => {
     const validationMessage = 'We were unable to contact you.Enter your email below.';
-    const forgotPasswordPage = mount(reduxWrapper(<IntlForgotPasswordPage {...props} />));
+    const { container } = render(reduxWrapper(<IntlForgotPasswordPage {...props} />));
 
-    await act(async () => { await forgotPasswordPage.find('button.btn-brand').simulate('click'); });
+    const submitButton = screen.getByText('Submit');
+    fireEvent.click(submitButton);
 
-    forgotPasswordPage.update();
-    expect(forgotPasswordPage.find('.alert-danger').text()).toEqual(validationMessage);
+    const alertElements = container.querySelectorAll('.alert-danger');
+    const validationErrors = alertElements[0].textContent;
+
+    expect(validationErrors).toBe(validationMessage);
   });
 
   it('should display request in progress error message', () => {
@@ -119,18 +146,22 @@ describe('ForgotPasswordPage', () => {
       forgotPassword: { status: 'forbidden' },
     });
 
-    const forgotPasswordPage = mount(reduxWrapper(<IntlForgotPasswordPage {...props} />));
-    expect(forgotPasswordPage.find('.alert-danger').text()).toEqual(rateLimitMessage);
+    const { container } = render(reduxWrapper(<IntlForgotPasswordPage {...props} />));
+
+    const alertElements = container.querySelectorAll('.alert-danger');
+    const validationErrors = alertElements[0].textContent;
+    expect(validationErrors).toBe(rateLimitMessage);
   });
 
   it('should not display any error message on change event', () => {
-    const forgotPasswordPage = mount(reduxWrapper(<IntlForgotPasswordPage {...props} />));
+    render(reduxWrapper(<IntlForgotPasswordPage {...props} />));
 
-    const emailInput = forgotPasswordPage.find('input#email');
-    emailInput.simulate('change', { target: { value: 'invalid-email', name: 'email' } });
-    forgotPasswordPage.update();
+    const emailInput = screen.getByLabelText('Email');
 
-    expect(forgotPasswordPage.find('#email-invalid-feedback').exists()).toEqual(false);
+    fireEvent.change(emailInput, { target: { value: 'invalid-email' } });
+
+    const errorElement = screen.queryByTestId('email-invalid-feedback');
+    expect(errorElement).toBeNull();
   });
 
   it('should set error in redux store on onBlur', () => {
@@ -146,8 +177,11 @@ describe('ForgotPasswordPage', () => {
     };
 
     store.dispatch = jest.fn(store.dispatch);
-    const forgotPasswordPage = mount(reduxWrapper(<IntlForgotPasswordPage {...props} />));
-    forgotPasswordPage.find('input#email').simulate('blur');
+    render(reduxWrapper(<IntlForgotPasswordPage {...props} />));
+    const emailInput = screen.getByLabelText('Email');
+
+    fireEvent.blur(emailInput);
+
     expect(store.dispatch).toHaveBeenCalledWith(setForgotPasswordFormData(forgotPasswordFormData));
   });
 
@@ -158,9 +192,9 @@ describe('ForgotPasswordPage', () => {
       emailValidationError: validationMessage,
       email: '',
     };
-    const forgotPasswordPage = mount(reduxWrapper(<IntlForgotPasswordPage {...props} />));
-    forgotPasswordPage.update();
-    expect(forgotPasswordPage.find('.pgn__form-text-invalid').text()).toEqual(validationMessage);
+    const { container } = render(reduxWrapper(<IntlForgotPasswordPage {...props} />));
+    const validationElement = container.querySelector('.pgn__form-text-invalid');
+    expect(validationElement.textContent).toEqual(validationMessage);
   });
 
   it('should clear error in redux store on onFocus', () => {
@@ -175,8 +209,12 @@ describe('ForgotPasswordPage', () => {
     };
 
     store.dispatch = jest.fn(store.dispatch);
-    const forgotPasswordPage = mount(reduxWrapper(<IntlForgotPasswordPage {...props} />));
-    forgotPasswordPage.find('input#email').simulate('focus');
+
+    render(reduxWrapper(<IntlForgotPasswordPage {...props} />));
+    const emailInput = screen.getByLabelText('Email');
+
+    fireEvent.focus(emailInput);
+
     expect(store.dispatch).toHaveBeenCalledWith(setForgotPasswordFormData(forgotPasswordFormData));
   });
 
@@ -186,14 +224,9 @@ describe('ForgotPasswordPage', () => {
       emailValidationError: '',
       email: '',
     };
-    const forgotPasswordPage = mount(reduxWrapper(<IntlForgotPasswordPage {...props} />));
-    forgotPasswordPage.update();
-    expect(forgotPasswordPage.find('#email-invalid-feedback').exists()).toEqual(false);
-  });
-
-  it('check cookie rendered', () => {
-    const forgotPage = mount(reduxWrapper(<IntlForgotPasswordPage {...props} />));
-    expect(forgotPage.find(<CookiePolicyBanner />)).toBeTruthy();
+    render(reduxWrapper(<IntlForgotPasswordPage {...props} />));
+    const errorElement = screen.queryByTestId('email-invalid-feedback');
+    expect(errorElement).toBeNull();
   });
 
   it('should display success message after email is sent', () => {
@@ -203,12 +236,16 @@ describe('ForgotPasswordPage', () => {
         status: 'complete',
       },
     });
+
     const successMessage = 'Check your emailWe sent an email to  with instructions to reset your password. If you do not '
                            + 'receive a password reset message after 1 minute, verify that you entered the correct email address,'
                            + ' or check your spam folder. If you need further assistance, contact technical support.';
 
-    const wrapper = mount(reduxWrapper(<IntlForgotPasswordPage {...props} />));
-    expect(wrapper.find('.alert-success').text()).toEqual(successMessage);
+    const { container } = render(reduxWrapper(<IntlForgotPasswordPage {...props} />));
+    const successElement = findByTextContent(container, successMessage);
+
+    expect(successElement).toBeDefined();
+    expect(successElement.textContent).toEqual(successMessage);
   });
 
   it('should display invalid password reset link error', () => {
@@ -222,20 +259,20 @@ describe('ForgotPasswordPage', () => {
                             + 'This password reset link is invalid. It may have been used already. '
                             + 'Enter your email below to receive a new link.';
 
-    const wrapper = mount(reduxWrapper(<IntlForgotPasswordPage {...props} />));
-    expect(wrapper.find('.alert-danger').text()).toEqual(successMessage);
+    const { container } = render(reduxWrapper(<IntlForgotPasswordPage {...props} />));
+    const successElement = findByTextContent(container, successMessage);
+
+    expect(successElement).toBeDefined();
+    expect(successElement.textContent).toEqual(successMessage);
   });
 
   it('should redirect onto login page', async () => {
-    const forgotPasswordPage = mount(reduxWrapper(
-      <Router history={history}>
-        <IntlForgotPasswordPage {...props} />
-      </Router>,
-    ));
+    const { container } = render(reduxWrapper(<IntlForgotPasswordPage {...props} />));
 
-    await act(async () => { await forgotPasswordPage.find('nav').find('a').first().simulate('click'); });
+    const navElement = container.querySelector('nav');
+    const anchorElement = navElement.querySelector('a');
+    fireEvent.click(anchorElement);
 
-    forgotPasswordPage.update();
-    expect(history.location.pathname).toEqual(LOGIN_PAGE);
+    expect(mockedNavigator).toHaveBeenCalledWith(LOGIN_PAGE);
   });
 });
