@@ -1,10 +1,10 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 
 import { getConfig } from '@edx/frontend-platform';
 import { getAuthService } from '@edx/frontend-platform/auth';
-import { FormattedMessage, injectIntl, intlShape } from '@edx/frontend-platform/i18n';
-import { Alert, Hyperlink } from '@edx/paragon';
-import { Error } from '@edx/paragon/icons';
+import { FormattedMessage, useIntl } from '@edx/frontend-platform/i18n';
+import { Alert, Hyperlink } from '@openedx/paragon';
+import { Error } from '@openedx/paragon/icons';
 import PropTypes from 'prop-types';
 
 import ChangePasswordPrompt from './ChangePasswordPrompt';
@@ -20,40 +20,55 @@ import {
   NON_COMPLIANT_PASSWORD_EXCEPTION,
   NUDGE_PASSWORD_CHANGE,
   REQUIRE_PASSWORD_CHANGE,
+  TPA_AUTHENTICATION_FAILURE,
 } from './data/constants';
 import messages from './messages';
+import { windowScrollTo } from '../data/utils';
 
 const LoginFailureMessage = (props) => {
-  const { intl } = props;
-  const { context, errorCode } = props.loginError;
+  const { formatMessage } = useIntl();
   const authService = getAuthService();
-  let errorList;
+  const {
+    context,
+    errorCode,
+    errorCount, // This is used to trigger the useEffect, facilitating the scrolling to the top.
+  } = props;
+
+  useEffect(() => {
+    windowScrollTo({ left: 0, top: 0, behavior: 'smooth' });
+  }, [errorCode, errorCount]);
+
+  if (!errorCode) {
+    return null;
+  }
+
   let resetLink = (
     <Hyperlink destination="reset" isInline>
-      {intl.formatMessage(messages['login.incorrect.credentials.error.reset.link.text'])}
+      {formatMessage(messages['login.incorrect.credentials.error.reset.link.text'])}
     </Hyperlink>
   );
 
+  let errorMessage;
   switch (errorCode) {
     case NON_COMPLIANT_PASSWORD_EXCEPTION: {
-      errorList = (
+      errorMessage = (
         <>
-          <strong>{intl.formatMessage(messages['non.compliant.password.title'])}</strong>
-          <p>{intl.formatMessage(messages['non.compliant.password.message'])}</p>
+          <strong>{formatMessage(messages['non.compliant.password.title'])}</strong>
+          <p>{formatMessage(messages['non.compliant.password.message'])}</p>
         </>
       );
       break;
     }
     case FORBIDDEN_REQUEST:
-      errorList = <p>{intl.formatMessage(messages['login.rate.limit.reached.message'])}</p>;
+      errorMessage = <p>{formatMessage(messages['login.rate.limit.reached.message'])}</p>;
       break;
     case INACTIVE_USER: {
       const supportLink = (
         <a href={context.supportLink}>
-          {intl.formatMessage(messages['contact.support.link'], { platformName: context.platformName })}
+          {formatMessage(messages['contact.support.link'], { platformName: context.platformName })}
         </a>
       );
-      errorList = (
+      errorMessage = (
         <p>
           <FormattedMessage
             id="login.inactive.user.error"
@@ -62,7 +77,7 @@ const LoginFailureMessage = (props) => {
             check your spam folders or {supportLink}."
             values={{
               lineBreak: <br />,
-              email: <strong className="data-hj-suppress">{props.loginError.email}</strong>,
+              email: <strong className="data-hj-suppress">{context.email}</strong>,
               supportLink,
             }}
           />
@@ -74,10 +89,10 @@ const LoginFailureMessage = (props) => {
       const url = `${getConfig().LMS_BASE_URL}/dashboard/?tpa_hint=${context.tpaHint}`;
       const tpaLink = (
         <a href={url}>
-          {intl.formatMessage(messages['tpa.account.link'], { provider: context.provider })}
+          {formatMessage(messages['tpa.account.link'], { provider: context.provider })}
         </a>
       );
-      errorList = (
+      errorMessage = (
         <p>
           <FormattedMessage
             id="allowed.domain.login.error"
@@ -90,15 +105,15 @@ const LoginFailureMessage = (props) => {
       break;
     }
     case INVALID_FORM:
-      errorList = <p>{intl.formatMessage(messages['login.form.invalid.error.message'])}</p>;
+      errorMessage = <p>{formatMessage(messages['login.form.invalid.error.message'])}</p>;
       break;
     case FAILED_LOGIN_ATTEMPT: {
       resetLink = (
         <Hyperlink destination="reset" isInline>
-          {intl.formatMessage(messages['login.incorrect.credentials.error.before.account.blocked.text'])}
+          {formatMessage(messages['login.incorrect.credentials.error.before.account.blocked.text'])}
         </Hyperlink>
       );
-      errorList = (
+      errorMessage = (
         <>
           <p>
             <FormattedMessage
@@ -122,9 +137,9 @@ const LoginFailureMessage = (props) => {
       break;
     }
     case ACCOUNT_LOCKED_OUT: {
-      errorList = (
+      errorMessage = (
         <>
-          <p>{intl.formatMessage(messages['account.locked.out.message.1'])}</p>
+          <p>{formatMessage(messages['account.locked.out.message.1'])}</p>
           <p>
             <FormattedMessage
               id="account.locked.out.message.2"
@@ -139,9 +154,9 @@ const LoginFailureMessage = (props) => {
     }
     case INCORRECT_EMAIL_PASSWORD:
       if (context.failureCount <= 1) {
-        errorList = <p>{intl.formatMessage(messages['login.incorrect.credentials.error'])}</p>;
+        errorMessage = <p>{formatMessage(messages['login.incorrect.credentials.error'])}</p>;
       } else if (context.failureCount === 2) {
-        errorList = (
+        errorMessage = (
           <p>
             <FormattedMessage
               id="login.incorrect.credentials.error.with.reset.link"
@@ -159,41 +174,56 @@ const LoginFailureMessage = (props) => {
       }
       return (
         <ChangePasswordPrompt
-          redirectUrl={props.loginError.redirectUrl}
+          redirectUrl={context.redirectUrl}
           variant="nudge"
         />
       );
     case REQUIRE_PASSWORD_CHANGE:
       return <ChangePasswordPrompt />;
+    case TPA_AUTHENTICATION_FAILURE:
+      errorMessage = (
+        <p>
+          {formatMessage(messages['login.tpa.authentication.failure'], {
+            platform_name: getConfig().SITE_NAME,
+            lineBreak: <br />,
+            errorMessage: context.errorMessage,
+          })}
+        </p>
+      );
+      break;
     case INTERNAL_SERVER_ERROR:
     default:
-      errorList = <p>{intl.formatMessage(messages['internal.server.error.message'])}</p>;
+      errorMessage = <p>{formatMessage(messages['internal.server.error.message'])}</p>;
       break;
   }
 
   return (
     <Alert id="login-failure-alert" className="mb-5" variant="danger" icon={Error}>
-      <Alert.Heading>{intl.formatMessage(messages['login.failure.header.title'])}</Alert.Heading>
-      { errorList }
+      <Alert.Heading>{formatMessage(messages['login.failure.header.title'])}</Alert.Heading>
+      { errorMessage }
     </Alert>
   );
 };
 
 LoginFailureMessage.defaultProps = {
-  loginError: {
-    redirectUrl: null,
-    errorCode: null,
-  },
+  context: {},
 };
 
 LoginFailureMessage.propTypes = {
-  loginError: PropTypes.shape({
-    context: PropTypes.object,
+  context: PropTypes.shape({
+    supportLink: PropTypes.string,
+    platformName: PropTypes.string,
+    tpaHint: PropTypes.string,
+    provider: PropTypes.string,
+    allowedDomain: PropTypes.string,
+    remainingAttempts: PropTypes.number,
+    failureCount: PropTypes.number,
+    errorMessage: PropTypes.string,
     email: PropTypes.string,
-    errorCode: PropTypes.string,
     redirectUrl: PropTypes.string,
   }),
-  intl: intlShape.isRequired,
+  errorCode: PropTypes.string.isRequired,
+  errorCount: PropTypes.number.isRequired,
 };
 
-export default injectIntl(LoginFailureMessage);
+export default LoginFailureMessage;
