@@ -1,64 +1,69 @@
+import {
+  BACKUP_LOGIN_DATA,
+  DISMISS_PASSWORD_RESET_BANNER,
+  LOGIN_REQUEST,
+} from './actions';
 import { DEFAULT_STATE, PENDING_STATE } from '../../data/constants';
 import { RESET_PASSWORD } from '../../reset-password';
-import { LOGIN_PERSIST_FORM_DATA, LOGIN_REMOVE_PASSWORD_RESET_BANNER, LOGIN_REQUEST } from './actions';
 
 export const defaultState = {
-  loginError: null,
+  loginErrorCode: '',
+  loginErrorContext: {},
   loginResult: {},
-  resetPassword: false,
   loginFormData: {
-    password: '',
-    emailOrUsername: '',
+    formFields: {
+      emailOrUsername: '', password: '',
+    },
     errors: {
-      emailOrUsername: '',
-      password: '',
+      emailOrUsername: '', password: '',
     },
   },
+  shouldBackupState: false,
+  showResetPasswordSuccessBanner: false,
+  submitState: DEFAULT_STATE,
 };
 
-const reducer = (state = defaultState, action) => {
+const reducer = (state = defaultState, action = {}) => {
   switch (action.type) {
+    case BACKUP_LOGIN_DATA.BASE:
+      return {
+        ...state,
+        shouldBackupState: true,
+      };
+    case BACKUP_LOGIN_DATA.BEGIN:
+      return {
+        ...defaultState,
+        loginFormData: { ...action.payload },
+      };
     case LOGIN_REQUEST.BEGIN:
       return {
         ...state,
+        showResetPasswordSuccessBanner: false,
         submitState: PENDING_STATE,
-        resetPassword: false,
       };
     case LOGIN_REQUEST.SUCCESS:
       return {
         ...state,
         loginResult: action.payload,
       };
-    case LOGIN_REQUEST.FAILURE:
+    case LOGIN_REQUEST.FAILURE: {
+      const { email, loginError, redirectUrl } = action.payload;
       return {
         ...state,
-        loginError: action.payload.loginError,
+        loginErrorCode: loginError.errorCode,
+        loginErrorContext: { ...loginError.context, email, redirectUrl },
         submitState: DEFAULT_STATE,
       };
-    case LOGIN_REQUEST.RESET:
-      return {
-        ...state,
-        loginError: null,
-      };
+    }
     case RESET_PASSWORD.SUCCESS:
       return {
         ...state,
-        resetPassword: true,
+        showResetPasswordSuccessBanner: true,
       };
-    case LOGIN_PERSIST_FORM_DATA: {
-      const { formData } = action.payload;
+    case DISMISS_PASSWORD_RESET_BANNER: {
       return {
         ...state,
-        loginFormData: {
-          ...state.loginFormData,
-          ...formData,
-        },
-      };
-    }
-    case LOGIN_REMOVE_PASSWORD_RESET_BANNER: {
-      return {
-        ...state,
-        resetPassword: false,
+        showResetPasswordSuccessBanner: false,
       };
     }
     default:
