@@ -1,4 +1,5 @@
+export const storeName = 'login';
+
 export { default as LoginPage } from './LoginPage';
 export { default as reducer } from './data/reducers';
 export { default as saga } from './data/sagas';
-export { storeName } from './data/selectors';
