@@ -2,7 +2,9 @@ import React from 'react';
 
 import { mergeConfig } from '@edx/frontend-platform';
 import { injectIntl, IntlProvider } from '@edx/frontend-platform/i18n';
-import { mount } from 'enzyme';
+import {
+  render, screen,
+} from '@testing-library/react';
 
 import AccountActivationMessage from '../AccountActivationMessage';
 import { ACCOUNT_ACTIVATION_MESSAGE } from '../data/constants';
@@ -17,18 +19,22 @@ describe('AccountActivationMessage', () => {
   });
 
   it('should match account already activated message', () => {
-    const accountActivationMessage = mount(
+    render(
       <IntlProvider locale="en">
         <IntlAccountActivationMessage messageType={ACCOUNT_ACTIVATION_MESSAGE.INFO} />
       </IntlProvider>,
     );
 
     const expectedMessage = 'This account has already been activated.';
-    expect(accountActivationMessage.find('#account-activation-message').find('div').first().text()).toEqual(expectedMessage);
+
+    expect(screen.getByText(
+      '',
+      { selector: '#account-activation-message' },
+    ).textContent).toBe(expectedMessage);
   });
 
   it('should match account activated success message', () => {
-    const accountActivationMessage = mount(
+    render(
       <IntlProvider locale="en">
         <IntlAccountActivationMessage messageType={ACCOUNT_ACTIVATION_MESSAGE.SUCCESS} />
       </IntlProvider>,
@@ -37,11 +43,15 @@ describe('AccountActivationMessage', () => {
     const expectedMessage = 'Success! You have activated your account.'
                             + 'You will now receive email updates and alerts from us related to '
                             + 'the courses you are enrolled in. Sign in to continue.';
-    expect(accountActivationMessage.find('#account-activation-message').first().text()).toEqual(expectedMessage);
+
+    expect(screen.getByText(
+      '',
+      { selector: '#account-activation-message' },
+    ).textContent).toBe(expectedMessage);
   });
 
   it('should match account activation error message', () => {
-    const accountActivationMessage = mount(
+    render(
       <IntlProvider locale="en">
         <IntlAccountActivationMessage messageType={ACCOUNT_ACTIVATION_MESSAGE.ERROR} />
       </IntlProvider>,
@@ -49,17 +59,22 @@ describe('AccountActivationMessage', () => {
 
     const expectedMessage = 'Your account could not be activated'
                             + 'Something went wrong, please contact support to resolve this issue.';
-    expect(accountActivationMessage.find('#account-activation-message').first().text()).toEqual(expectedMessage);
+
+    expect(screen.getByText(
+      '',
+      { selector: '#account-activation-message' },
+    ).textContent).toBe(expectedMessage);
   });
 
   it('should not display anything for invalid message type', () => {
-    const accountActivationMessage = mount(
+    const { container } = render(
       <IntlProvider locale="en">
         <IntlAccountActivationMessage messageType="invalid-message" />
       </IntlProvider>,
     );
 
-    expect(accountActivationMessage).toEqual({});
+    const accountActivationMessage = container.querySelectorAll('#account-activation-message');
+    expect(accountActivationMessage[0]).toBe(undefined);
   });
 });
 
@@ -71,36 +86,45 @@ describe('EmailConfirmationMessage', () => {
   });
 
   it('should match email already confirmed message', () => {
-    const accountVerificationMessage = mount(
+    render(
       <IntlProvider locale="en">
         <IntlAccountActivationMessage messageType={ACCOUNT_ACTIVATION_MESSAGE.INFO} />
       </IntlProvider>,
     );
 
     const expectedMessage = 'This email has already been confirmed.';
-    expect(accountVerificationMessage.find('#account-activation-message').find('div').first().text()).toEqual(expectedMessage);
+
+    expect(screen.getByText(
+      '',
+      { selector: '#account-activation-message' },
+    ).textContent).toBe(expectedMessage);
   });
 
   it('should match email confirmation success message', () => {
-    const accountVerificationMessage = mount(
+    render(
       <IntlProvider locale="en">
         <IntlAccountActivationMessage messageType={ACCOUNT_ACTIVATION_MESSAGE.SUCCESS} />
       </IntlProvider>,
     );
-
     const expectedMessage = 'Success! You have confirmed your email.Sign in to continue.';
-    expect(accountVerificationMessage.find('#account-activation-message').first().text()).toEqual(expectedMessage);
+
+    expect(screen.getByText(
+      '',
+      { selector: '#account-activation-message' },
+    ).textContent).toBe(expectedMessage);
   });
 
   it('should match email confirmation error message', () => {
-    const accountVerificationMessage = mount(
+    render(
       <IntlProvider locale="en">
         <IntlAccountActivationMessage messageType={ACCOUNT_ACTIVATION_MESSAGE.ERROR} />
       </IntlProvider>,
     );
-
     const expectedMessage = 'Your email could not be confirmed'
                             + 'Something went wrong, please contact support to resolve this issue.';
-    expect(accountVerificationMessage.find('#account-activation-message').first().text()).toEqual(expectedMessage);
+    expect(screen.getByText(
+      '',
+      { selector: '#account-activation-message' },
+    ).textContent).toBe(expectedMessage);
   });
 });
