@@ -1,11 +1,13 @@
 import React from 'react';
 
-import * as auth from '@edx/frontend-platform/auth';
 import { injectIntl, IntlProvider } from '@edx/frontend-platform/i18n';
-import { mount } from 'enzyme';
+import {
+  render, screen,
+} from '@testing-library/react';
 import { MemoryRouter } from 'react-router-dom';
 
 import {
+  ACCOUNT_LOCKED_OUT,
   ALLOWED_DOMAIN_LOGIN_ERROR,
   FAILED_LOGIN_ATTEMPT,
   FORBIDDEN_REQUEST,
@@ -16,11 +18,13 @@ import {
   NON_COMPLIANT_PASSWORD_EXCEPTION,
   NUDGE_PASSWORD_CHANGE,
   REQUIRE_PASSWORD_CHANGE,
+  TPA_AUTHENTICATION_FAILURE,
 } from '../data/constants';
 import LoginFailureMessage from '../LoginFailure';
 
-jest.mock('@edx/frontend-platform/auth');
-auth.getAuthService = jest.fn();
+jest.mock('@edx/frontend-platform/auth', () => ({
+  getAuthService: jest.fn(),
+}));
 
 const IntlLoginFailureMessage = injectIntl(LoginFailureMessage);
 
@@ -38,12 +42,11 @@ describe('LoginFailureMessage', () => {
 
   it('should match non compliant password error message', () => {
     props = {
-      loginError: {
-        errorCode: NON_COMPLIANT_PASSWORD_EXCEPTION,
-      },
+      errorCode: NON_COMPLIANT_PASSWORD_EXCEPTION,
+      errorCount: 0,
     };
 
-    const loginFailureMessage = mount(
+    render(
       <IntlProvider locale="en">
         <IntlLoginFailureMessage {...props} />
       </IntlProvider>,
@@ -54,22 +57,24 @@ describe('LoginFailureMessage', () => {
                             + 'password-reset message to the email address associated with this account. '
                             + 'Thank you for helping us keep your data safe.';
 
-    expect(loginFailureMessage.find('#login-failure-alert').first().text()).toEqual(expectedMessage);
+    expect(screen.getByText(
+      '',
+      { selector: '#login-failure-alert' },
+    ).textContent).toBe(expectedMessage);
   });
 
   it('should match inactive user error message', () => {
     props = {
-      loginError: {
+      context: {
         email: 'text@example.com',
-        errorCode: INACTIVE_USER,
-        context: {
-          platformName: 'openedX',
-          supportLink: 'https://support.edx.org/',
-        },
+        platformName: 'openedX',
+        supportLink: 'http://support.openedx.test',
       },
+      errorCode: INACTIVE_USER,
+      errorCount: 0,
     };
 
-    const loginFailureMessage = mount(
+    render(
       <IntlProvider locale="en">
         <IntlLoginFailureMessage {...props} />
       </IntlProvider>,
@@ -79,154 +84,221 @@ describe('LoginFailureMessage', () => {
                             + 'We just sent an activation link to text@example.com. If you do not receive an email, '
                             + 'check your spam folders or contact openedX support.';
 
-    expect(loginFailureMessage.find('#login-failure-alert').first().text()).toEqual(expectedMessage);
-    expect(loginFailureMessage.find('#login-failure-alert').find('a').props().href).toEqual('https://support.edx.org/');
+    expect(screen.getByText(
+      '',
+      { selector: '#login-failure-alert' },
+    ).textContent).toBe(expectedMessage);
+
+    expect(screen.getByRole('link', { name: 'contact openedX support' }).getAttribute('href')).toBe('http://support.openedx.test');
   });
 
   it('test match failed login attempt error', () => {
     props = {
-      loginError: {
+      context: {
         email: 'text@example.com',
-        errorCode: FAILED_LOGIN_ATTEMPT,
-        context: {
-          remainingAttempts: 3,
-          allowedFailureAttempts: 6,
-          resetLink: '/reset',
-        },
+        remainingAttempts: 3,
+        allowedFailureAttempts: 6,
+        resetLink: '/reset',
       },
+      errorCode: FAILED_LOGIN_ATTEMPT,
+      errorCount: 0,
     };
 
-    const loginFailureMessage = mount(
+    render(
       <IntlProvider locale="en">
         <IntlLoginFailureMessage {...props} />
       </IntlProvider>,
     );
+
     const expectedMessage = 'We couldn\'t sign you in.The username, email or password you entered is incorrect. '
                             + 'You have 3 more sign in attempts before your account is temporarily locked.If you\'ve forgotten your password, click here to reset it.';
 
-    expect(loginFailureMessage.find('#login-failure-alert').first().text()).toEqual(expectedMessage);
+    expect(screen.getByText(
+      '',
+      { selector: '#login-failure-alert' },
+    ).textContent).toBe(expectedMessage);
   });
 
   it('test match failed login error first attempt', () => {
     props = {
-      loginError: {
+      context: {
         email: 'text@example.com',
-        errorCode: INCORRECT_EMAIL_PASSWORD,
-        context: {
-          failureCount: 1,
-          resetLink: '/reset',
-        },
+        failureCount: 1,
+        resetLink: '/reset',
       },
+      errorCode: INCORRECT_EMAIL_PASSWORD,
+      errorCount: 0,
     };
 
-    const loginFailureMessage = mount(
+    render(
       <IntlProvider locale="en">
         <IntlLoginFailureMessage {...props} />
       </IntlProvider>,
     );
+
     const expectedMessage = 'We couldn\'t sign you in.The username, email, or password you entered is incorrect. Please try again.';
 
-    expect(loginFailureMessage.find('#login-failure-alert').first().text()).toEqual(expectedMessage);
+    expect(screen.getByText(
+      '',
+      { selector: '#login-failure-alert' },
+    ).textContent).toBe(expectedMessage);
+  });
+
+  it('test match user account locked out', () => {
+    props = {
+      errorCode: ACCOUNT_LOCKED_OUT,
+      errorCount: 0,
+    };
+
+    render(
+      <IntlProvider locale="en">
+        <IntlLoginFailureMessage {...props} />
+      </IntlProvider>,
+    );
+
+    const expectedMessage = 'We couldn\'t sign you in.To protect your account, it\'s been temporarily locked. Try again in 30 minutes.To be on the safe side, you can reset your password before trying again.';
+    expect(screen.getByText(
+      '',
+      { selector: '#login-failure-alert' },
+    ).textContent).toBe(expectedMessage);
   });
 
   it('test match failed login error second attempt', () => {
     props = {
-      loginError: {
+      context: {
         email: 'text@example.com',
-        errorCode: INCORRECT_EMAIL_PASSWORD,
-        context: {
-          failureCount: 2,
-          resetLink: '/reset',
-        },
+        failureCount: 2,
+        resetLink: '/reset',
       },
+      errorCode: INCORRECT_EMAIL_PASSWORD,
+      errorCount: 0,
     };
 
-    const loginFailureMessage = mount(
+    render(
       <IntlProvider locale="en">
         <IntlLoginFailureMessage {...props} />
       </IntlProvider>,
     );
+
     const expectedMessage = 'We couldn\'t sign you in.The username, email, or password you entered is incorrect. Please try again or reset your password.';
 
-    expect(loginFailureMessage.find('#login-failure-alert').first().text()).toEqual(expectedMessage);
+    expect(screen.getByText(
+      '',
+      { selector: '#login-failure-alert' },
+    ).textContent).toBe(expectedMessage);
   });
 
   it('should match rate limit error message', () => {
     props = {
-      loginError: {
-        errorCode: FORBIDDEN_REQUEST,
-      },
+      errorCode: FORBIDDEN_REQUEST,
+      errorCount: 0,
     };
 
-    const loginFailureMessage = mount(
+    render(
       <IntlProvider locale="en">
         <IntlLoginFailureMessage {...props} />
       </IntlProvider>,
     );
 
     const expectedMessage = 'We couldn\'t sign you in.Too many failed login attempts. Try again later.';
-    expect(loginFailureMessage.find('#login-failure-alert').first().text()).toEqual(expectedMessage);
+
+    expect(screen.getByText(
+      '',
+      { selector: '#login-failure-alert' },
+    ).textContent).toBe(expectedMessage);
   });
 
   it('should match internal server error message', () => {
     props = {
-      loginError: {
-        errorCode: INTERNAL_SERVER_ERROR,
-      },
+      errorCode: INTERNAL_SERVER_ERROR,
+      errorCount: 0,
     };
 
-    const loginFailureMessage = mount(
+    render(
       <IntlProvider locale="en">
         <IntlLoginFailureMessage {...props} />
       </IntlProvider>,
     );
 
     const expectedMessage = 'We couldn\'t sign you in.An error has occurred. Try refreshing the page, or check your internet connection.';
-    expect(loginFailureMessage.find('#login-failure-alert').first().text()).toEqual(expectedMessage);
+
+    expect(screen.getByText(
+      '',
+      { selector: '#login-failure-alert' },
+    ).textContent).toBe(expectedMessage);
   });
 
   it('should match invalid form error message', () => {
     props = {
-      loginError: {
-        errorCode: INVALID_FORM,
-      },
+      errorCode: INVALID_FORM,
+      errorCount: 0,
     };
 
-    const loginFailureMessage = mount(
+    render(
       <IntlProvider locale="en">
         <IntlLoginFailureMessage {...props} />
       </IntlProvider>,
     );
 
     const expectedMessage = 'We couldn\'t sign you in.Please fill in the fields below.';
-    expect(loginFailureMessage.find('#login-failure-alert').first().text()).toEqual(expectedMessage);
+    expect(screen.getByText(
+      '',
+      { selector: '#login-failure-alert' },
+    ).textContent).toBe(expectedMessage);
   });
 
   it('should match internal server of error message', () => {
     props = {
-      loginError: {
-        errorCode: 'invalid-error-code',
-      },
+      errorCode: 'invalid-error-code',
+      errorCount: 0,
     };
 
-    const loginFailureMessage = mount(
+    render(
       <IntlProvider locale="en">
         <IntlLoginFailureMessage {...props} />
       </IntlProvider>,
     );
 
     const expectedMessage = 'We couldn\'t sign you in.An error has occurred. Try refreshing the page, or check your internet connection.';
-    expect(loginFailureMessage.find('#login-failure-alert').first().text()).toEqual(expectedMessage);
+    expect(screen.getByText(
+      '',
+      { selector: '#login-failure-alert' },
+    ).textContent).toBe(expectedMessage);
+  });
+
+  it('should match tpa authentication failed error message', () => {
+    props = {
+      errorCode: TPA_AUTHENTICATION_FAILURE,
+      errorCount: 0,
+      context: { errorMessage: 'An error occurred' },
+    };
+
+    render(
+      <IntlProvider locale="en">
+        <IntlLoginFailureMessage {...props} />
+      </IntlProvider>,
+    );
+
+    const expectedMessageSubstring = 'We are sorry, you are not authorized to access';
+
+    expect(screen.getByText(
+      '',
+      { selector: '#login-failure-alert' },
+    ).textContent).toContain(expectedMessageSubstring);
+
+    expect(screen.getByText(
+      '',
+      { selector: '#login-failure-alert' },
+    ).textContent).toContain('An error occurred');
   });
 
   it('should show modal that nudges users to change password', () => {
     props = {
-      loginError: {
-        errorCode: NUDGE_PASSWORD_CHANGE,
-      },
+      errorCode: NUDGE_PASSWORD_CHANGE,
+      errorCount: 0,
     };
 
-    const loginFailureMessage = mount(
+    render(
       <IntlProvider locale="en">
         <MemoryRouter>
           <IntlLoginFailureMessage {...props} />
@@ -234,21 +306,25 @@ describe('LoginFailureMessage', () => {
       </IntlProvider>,
     );
 
-    expect(loginFailureMessage.find('.pgn__modal-title').text()).toEqual('Password security');
-    expect(loginFailureMessage.find('.pgn__modal-body').text()).toEqual(
-      'Our system detected that your password is vulnerable. '
-               + 'We recommend you change it so that your account stays secure.',
-    );
+    const message = 'Our system detected that your password is vulnerable. '
+                         + 'We recommend you change it so that your account stays secure.';
+    expect(screen.getByText(
+      'Password security',
+      { selector: '.pgn__modal-title' },
+    ).textContent).toEqual('Password security');
+    expect(screen.getByText(
+      '',
+      { selector: '.pgn__modal-body' },
+    ).textContent).toEqual(message);
   });
 
   it('should show modal that requires users to change password', () => {
     props = {
-      loginError: {
-        errorCode: REQUIRE_PASSWORD_CHANGE,
-      },
+      errorCode: REQUIRE_PASSWORD_CHANGE,
+      errorCount: 0,
     };
 
-    const loginFailureMessage = mount(
+    render(
       <IntlProvider locale="en">
         <MemoryRouter>
           <IntlLoginFailureMessage {...props} />
@@ -256,8 +332,14 @@ describe('LoginFailureMessage', () => {
       </IntlProvider>,
     );
 
-    expect(loginFailureMessage.find('.pgn__modal-title').text()).toEqual('Password change required');
-    expect(loginFailureMessage.find('.pgn__modal-body').text()).toEqual(
+    expect(screen.getByText(
+      'Password change required',
+      { selector: '.pgn__modal-title' },
+    ).textContent).toEqual('Password change required');
+    expect(screen.getByText(
+      '',
+      { selector: '.pgn__modal-body' },
+    ).textContent).toEqual(
       'Our system detected that your password is vulnerable. '
                + 'Change your password so that your account stays secure.',
     );
@@ -265,27 +347,30 @@ describe('LoginFailureMessage', () => {
 
   it('should show message if staff user try to login through password', () => {
     props = {
-      loginError: {
+      context: {
         email: 'text@example.com',
-        errorCode: ALLOWED_DOMAIN_LOGIN_ERROR,
-        context: {
-          allowedDomain: 'edx.org',
-          provider: 'Google',
-          tpaHint: 'google-auth2',
-        },
+        allowedDomain: 'test.com',
+        provider: 'Google',
+        tpaHint: 'google-auth2',
       },
+      errorCode: ALLOWED_DOMAIN_LOGIN_ERROR,
+      errorCount: 0,
     };
 
-    const loginFailureMessage = mount(
+    render(
       <IntlProvider locale="en">
         <IntlLoginFailureMessage {...props} />
       </IntlProvider>,
     );
 
-    const errorMessage = "We couldn't sign you in.As edx.org user, You must login with your edx.org Google account.";
+    const errorMessage = "We couldn't sign you in.As test.com user, You must login with your test.com Google account.";
     const url = 'http://localhost:18000/dashboard/?tpa_hint=google-auth2';
 
-    expect(loginFailureMessage.find('#login-failure-alert').first().text()).toEqual(errorMessage);
-    expect(loginFailureMessage.find('#login-failure-alert').find('a').props().href).toEqual(url);
+    expect(screen.getByText(
+      '',
+      { selector: '#login-failure-alert' },
+    ).textContent).toContain(errorMessage);
+
+    expect(screen.getByRole('link', { name: 'Google account' }).getAttribute('href')).toBe(url);
   });
 });
