@@ -1,4 +1,4 @@
 export { default as RegistrationPage } from './RegistrationPage';
 export { default as reducer } from './data/reducers';
 export { default as saga } from './data/sagas';
-export { storeName } from './data/selectors';
+export { storeName } from './data/reducers';
