@@ -1,31 +1,32 @@
 import React from 'react';
 
-import { injectIntl, intlShape } from '@edx/frontend-platform/i18n';
-import { Alert } from '@edx/paragon';
-import { Error } from '@edx/paragon/icons';
+import { useIntl } from '@edx/frontend-platform/i18n';
+import { Alert } from '@openedx/paragon';
+import { Error } from '@openedx/paragon/icons';
 import PropTypes from 'prop-types';
 
 import { FORM_SUBMISSION_ERROR, PASSWORD_RESET, PASSWORD_VALIDATION_ERROR } from './data/constants';
 import messages from './messages';
 
 const ResetPasswordFailure = (props) => {
-  const { errorCode, errorMsg, intl } = props;
+  const { formatMessage } = useIntl();
+  const { errorCode, errorMsg } = props;
 
   let errorMessage = null;
-  let heading = intl.formatMessage(messages['reset.password.failure.heading']);
+  let heading = formatMessage(messages['reset.password.failure.heading']);
   switch (errorCode) {
     case PASSWORD_RESET.FORBIDDEN_REQUEST:
-      heading = intl.formatMessage(messages['reset.server.rate.limit.error']);
-      errorMessage = intl.formatMessage(messages['rate.limit.error']);
+      heading = formatMessage(messages['reset.server.rate.limit.error']);
+      errorMessage = formatMessage(messages['rate.limit.error']);
       break;
     case PASSWORD_RESET.INTERNAL_SERVER_ERROR:
-      errorMessage = intl.formatMessage(messages['internal.server.error']);
+      errorMessage = formatMessage(messages['internal.server.error']);
       break;
     case PASSWORD_VALIDATION_ERROR:
       errorMessage = errorMsg;
      break;
     case FORM_SUBMISSION_ERROR:
-      errorMessage = intl.formatMessage(messages['reset.password.form.submission.error']);
+      errorMessage = formatMessage(messages['reset.password.form.submission.error']);
       break;
     default:
       break;
@@ -51,7 +52,6 @@ ResetPasswordFailure.defaultProps = {
 ResetPasswordFailure.propTypes = {
   errorCode: PropTypes.string,
   errorMsg: PropTypes.string,
-  intl: intlShape.isRequired,
 };
 
-export default injectIntl(ResetPasswordFailure);
+export default ResetPasswordFailure;
