@@ -1,8 +1,8 @@
 import React, { useEffect, useState } from 'react';
 import { connect } from 'react-redux';
 
-import { getConfig, getQueryParameters } from '@edx/frontend-platform';
-import { injectIntl, intlShape } from '@edx/frontend-platform/i18n';
+import { getConfig } from '@edx/frontend-platform';
+import { useIntl } from '@edx/frontend-platform/i18n';
 import {
   Form,
   Icon,
@@ -10,18 +10,12 @@ import {
   StatefulButton,
   Tab,
   Tabs,
-} from '@edx/paragon';
-import { ChevronLeft } from '@edx/paragon/icons';
+} from '@openedx/paragon';
+import { ChevronLeft } from '@openedx/paragon/icons';
 import PropTypes from 'prop-types';
 import { Helmet } from 'react-helmet';
-import { Redirect } from 'react-router-dom';
+import { useNavigate, useParams } from 'react-router-dom';
 
-import BaseComponent from '../base-component';
-import { PasswordField } from '../common-components';
-import {
-  LETTER_REGEX, LOGIN_PAGE, NUMBER_REGEX, RESET_PAGE,
-} from '../data/constants';
-import { updatePathWithQueryParams, windowScrollTo } from '../data/utils';
 import { resetPassword, validateToken } from './data/actions';
 import {
   FORM_SUBMISSION_ERROR, PASSWORD_RESET_ERROR, PASSWORD_VALIDATION_ERROR, TOKEN_STATE,
@@ -30,24 +24,32 @@ import { resetPasswordResultSelector } from './data/selectors';
 import { validatePassword } from './data/service';
 import messages from './messages';
 import ResetPasswordFailure from './ResetPasswordFailure';
+import BaseContainer from '../base-container';
+import { PasswordField } from '../common-components';
+import {
+  LETTER_REGEX, LOGIN_PAGE, NUMBER_REGEX, RESET_PAGE,
+} from '../data/constants';
+import { getAllPossibleQueryParams, updatePathWithQueryParams, windowScrollTo } from '../data/utils';
 
 const ResetPasswordPage = (props) => {
-  const { intl } = props;
+  const { formatMessage } = useIntl();
+  const newPasswordError = formatMessage(messages['password.validation.message']);
 
   const [newPassword, setNewPassword] = useState('');
   const [confirmPassword, setConfirmPassword] = useState('');
   const [formErrors, setFormErrors] = useState({});
   const [errorCode, setErrorCode] = useState(null);
-  const [key, setKey] = useState('');
+  const { token } = useParams();
+  const navigate = useNavigate();
 
   useEffect(() => {
     if (props.status !== TOKEN_STATE.PENDING && props.status !== PASSWORD_RESET_ERROR) {
       setErrorCode(props.status);
     }
     if (props.status === PASSWORD_VALIDATION_ERROR) {
-      setFormErrors({ newPassword: intl.formatMessage(messages['password.validation.message']) });
+      setFormErrors({ newPassword: newPasswordError });
     }
-  }, [props.status, intl]);
+  }, [props.status, newPasswordError]);
 
   const validatePasswordFromBackend = async (password) => {
     let errorMessage = '';
@@ -67,16 +69,16 @@ const ResetPasswordPage = (props) => {
     switch (name) {
       case 'newPassword':
         if (!value || !LETTER_REGEX.test(value) || !NUMBER_REGEX.test(value) || value.length < 8) {
-          formErrors.newPassword = intl.formatMessage(messages['password.validation.message']);
+          formErrors.newPassword = formatMessage(messages['password.validation.message']);
         } else {
           validatePasswordFromBackend(value);
         }
         break;
       case 'confirmPassword':
         if (!value) {
-          formErrors.confirmPassword = intl.formatMessage(messages['confirm.your.password']);
+          formErrors.confirmPassword = formatMessage(messages['confirm.your.password']);
         } else if (value !== newPassword) {
-          formErrors.confirmPassword = intl.formatMessage(messages['passwords.do.not.match']);
+          formErrors.confirmPassword = formatMessage(messages['passwords.do.not.match']);
         } else {
           formErrors.confirmPassword = '';
         }
@@ -89,20 +91,7 @@ const ResetPasswordPage = (props) => {
   };
 
   const handleOnBlur = (event) => {
-    let { name, value } = event.target;
-
-    // Do not validate when focus out from 'newPassword' and focus on 'passwordValidation' icon
-    // for better user experience.
-    if (event.relatedTarget
-      && event.relatedTarget.name === 'passwordValidation'
-      && name === 'newPassword'
-    ) {
-      return;
-    }
-    if (name === 'passwordValidation') {
-      name = 'newPassword';
-      value = newPassword;
-    }
+    const { name, value } = event.target;
     validateInput(name, value);
   };
 
@@ -128,7 +117,7 @@ const ResetPasswordPage = (props) => {
         new_password1: newPassword,
         new_password2: confirmPassword,
       };
-      const params = getQueryParameters();
+      const params = getAllPossibleQueryParams();
       props.resetPassword(formPayload, props.token, params);
     } else {
       setErrorCode(FORM_SUBMISSION_ERROR);
@@ -139,40 +128,36 @@ const ResetPasswordPage = (props) => {
   const tabTitle = (
     <div className="d-inline-flex flex-wrap align-items-center">
       <Icon src={ChevronLeft} />
-      <span className="ml-2">{intl.formatMessage(messages['sign.in'])}</span>
+      <span className="ml-2">{formatMessage(messages['sign.in'])}</span>
     </div>
   );
 
   if (props.status === TOKEN_STATE.PENDING) {
-    const { token } = props.match.params;
     if (token) {
       props.validateToken(token);
-      return <Spinner animation="border" variant="primary" className="centered-align-spinner" />;
+      return <Spinner animation="border" variant="primary" className="spinner--position-centered" />;
     }
   } else if (props.status === PASSWORD_RESET_ERROR) {
-    return <Redirect to={updatePathWithQueryParams(RESET_PAGE)} />;
+    navigate(updatePathWithQueryParams(RESET_PAGE));
   } else if (props.status === 'success') {
-    return <Redirect to={updatePathWithQueryParams(LOGIN_PAGE)} />;
+    navigate(updatePathWithQueryParams(LOGIN_PAGE));
   } else {
     return (
-      <BaseComponent>
+      <BaseContainer>
         <div>
           <Helmet>
-            <title>{intl.formatMessage(messages['reset.password.page.title'],
-              { siteName: getConfig().SITE_NAME })}
+            <title>
+              {formatMessage(messages['reset.password.page.title'], { siteName: getConfig().SITE_NAME })}
             </title>
           </Helmet>
-          <Tabs activeKey="" id="controlled-tab" onSelect={(k) => setKey(k)}>
+          <Tabs activeKey="" id="controlled-tab" onSelect={(key) => navigate(updatePathWithQueryParams(key))}>
             <Tab title={tabTitle} eventKey={LOGIN_PAGE} />
           </Tabs>
-          { key && (
-            <Redirect to={updatePathWithQueryParams(key)} />
-          )}
           <div id="main-content" className="main-content">
             <div className="mw-xs">
               <ResetPasswordFailure errorCode={errorCode} errorMsg={props.errorMsg} />
-              <h4>{intl.formatMessage(messages['reset.password'])}</h4>
-              <p className="mb-4">{intl.formatMessage(messages['reset.password.page.instructions'])}</p>
+              <h4>{formatMessage(messages['reset.password'])}</h4>
+              <p className="mb-4">{formatMessage(messages['reset.password.page.instructions'])}</p>
               <Form id="set-reset-password-form" name="set-reset-password-form">
                 <PasswordField
                   name="newPassword"
@@ -181,7 +166,7 @@ const ResetPasswordPage = (props) => {
                   handleBlur={handleOnBlur}
                   handleFocus={handleOnFocus}
                   errorMessage={formErrors.newPassword}
-                  floatingLabel={intl.formatMessage(messages['new.password.label'])}
+                  floatingLabel={formatMessage(messages['new.password.label'])}
                 />
                 <PasswordField
                   name="confirmPassword"
@@ -190,17 +175,17 @@ const ResetPasswordPage = (props) => {
                   handleFocus={handleOnFocus}
                   errorMessage={formErrors.confirmPassword}
                   showRequirements={false}
-                  floatingLabel={intl.formatMessage(messages['confirm.password.label'])}
+                  floatingLabel={formatMessage(messages['confirm.password.label'])}
                 />
                 <StatefulButton
                   id="submit-new-password"
                   name="submit-new-password"
                   type="submit"
                   variant="brand"
-                  className="stateful-button-width"
+                  className="reset-password--button"
                   state={props.status}
                   labels={{
-                    default: intl.formatMessage(messages['reset.password']),
+                    default: formatMessage(messages['reset.password']),
                     pending: '',
                   }}
                   onClick={e => handleSubmit(e)}
@@ -210,7 +195,7 @@ const ResetPasswordPage = (props) => {
             </div>
           </div>
         </div>
-      </BaseComponent>
+      </BaseContainer>
     );
   }
   return null;
@@ -219,20 +204,13 @@ const ResetPasswordPage = (props) => {
 ResetPasswordPage.defaultProps = {
   status: null,
   token: null,
-  match: null,
   errorMsg: null,
 };
 
 ResetPasswordPage.propTypes = {
-  intl: intlShape.isRequired,
   resetPassword: PropTypes.func.isRequired,
   validateToken: PropTypes.func.isRequired,
   token: PropTypes.string,
-  match: PropTypes.shape({
-    params: PropTypes.shape({
-      token: PropTypes.string,
-    }),
-  }),
   status: PropTypes.string,
   errorMsg: PropTypes.string,
 };
@@ -243,4 +221,4 @@ export default connect(
     resetPassword,
     validateToken,
   },
-)(injectIntl(ResetPasswordPage));
+)(ResetPasswordPage);
