@@ -1,25 +1,21 @@
 import React from 'react';
 
-import { injectIntl, intlShape } from '@edx/frontend-platform/i18n';
-import { Alert } from '@edx/paragon';
+import { useIntl } from '@edx/frontend-platform/i18n';
+import { Alert } from '@openedx/paragon';
 
 import messages from './messages';
 
-const ResetPasswordSuccess = (props) => {
-  const { intl } = props;
+const ResetPasswordSuccess = () => {
+  const { formatMessage } = useIntl();
 
   return (
-    <Alert id="reset-password-success" variant="success" className="mb-4">
+    <Alert id="reset-password-success" variant="success" className="mb-5">
       <Alert.Heading>
-        {intl.formatMessage(messages['reset.password.success.heading'])}
+        {formatMessage(messages['reset.password.success.heading'])}
       </Alert.Heading>
-      <p>{intl.formatMessage(messages['reset.password.success'])}</p>
+      <p>{formatMessage(messages['reset.password.success'])}</p>
     </Alert>
   );
 };
 
-ResetPasswordSuccess.propTypes = {
-  intl: intlShape.isRequired,
-};
-
-export default injectIntl(ResetPasswordSuccess);
+export default ResetPasswordSuccess;
