@@ -1,4 +1,4 @@
-export { default } from './ResetPasswordPage';
+export { default as ResetPasswordPage } from './ResetPasswordPage';
 export { default as reducer } from './data/reducers';
 export { RESET_PASSWORD } from './data/actions';
 export { default as saga } from './data/sagas';
